mgr/cephadm: test: Move test_monitoring_ports to TestMonitoring
